requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-ckeditor==6.1.0
djongo==1.3.6
pymongo==3.12.3
redis==3.5.3
googletrans==3.1.0a0
pytest==6.2.5
pytest-django==4.4.0
flake8==3.9.2

